Key synchronized implementation of TSpace
Locking and synchronizing on entire TSpace when inserting an item into
the queue has a negative impact on the liquidity of the work of other
threads that they try to insert elements to other queues. Which in turn
has an impact on performance of the whole system.
Therefore, as a container for the queues used ConcurrentHashMap instead
the synchronized HashMap.
ConcurrentHashMap allows without locking the simultaneous reading for
multiple threads, iterating through the map. Only write operations are
synchronized by Lock. It also allows the simultaneous iteration and
write the map without throwing ConcurrentModificationException.

Now the method notyfiListeners is synchronized with the insertion into
the queue.

The measured performance is about 2 to 3 times higher than the current
TSpace. The measured performance is about 2 to 3 times higher than the
current TSpace. Of course, a significant impact is dependent on the
number of CPU threads

Signed-off-by: Robert Demski <drdemsey@gmail.com>